LF-4158 update Origin component
